+++ src/EventStore.Core.Tests/Http/BasicAuthentication/basic_authentication.cs
@@ -10,7 +10,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 	namespace basic_authentication {
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_an_unprotected_resource<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override Task Given() => Task.CompletedTask;
 			protected override async Task When() {
@@ -31,7 +31,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_a_protected_resource<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override Task Given() => Task.CompletedTask;
 
@@ -53,7 +53,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_a_protected_resource_with_credentials_provided<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override async Task Given() {
 				var response = await MakeJsonPost(
@@ -73,7 +73,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_a_protected_resource_with_invalid_credentials_provided<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override async Task Given() {
 				var response = await MakeJsonPost(
@@ -93,7 +93,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_a_protected_resource_with_credentials_of_disabled_user_account<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override async Task Given() {
 				var response = await MakeJsonPost(
@@ -115,7 +115,7 @@ namespace EventStore.Core.Tests.Http.BasicAuthentication {
 
 		[Category("LongRunning")]
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_requesting_a_protected_resource_with_credentials_of_deleted_user_account<TLogFormat, TStreamId> : with_admin_user<TLogFormat, TStreamId> {
 			protected override async Task Given() {
 				var response = await MakeRawJsonPost(
